tests/ImazenShared.Tests/Abstractions/Concurrency/NonOverlappingAsyncRunnerTests.cs
namespace Imazen.Routing.Caching.Health.Tests;

using Xunit;
using System;
using System.Threading;
using System.Threading.Tasks;
using Imazen.Routing.Caching.Health;

public class NonOverlappingAsyncRunnerTests
{
    private class ArbitraryException : Exception { }
    private async ValueTask<int> TestTask30(CancellationToken ct)
    {
        await Task.Delay(30, ct);
        return 1;
    }
    private async ValueTask<int> TestTask80(CancellationToken ct)
    {
        await Task.Delay(80, ct);
        return 1;
    }

    [Fact]
    public async Task RunNonOverlappingAsync_ShouldStartNewTask_WhenNoTaskIsRunning()
    {
        var runner = new NonOverlappingAsyncRunner<int>(TestTask30);
        var result = await runner.RunNonOverlappingAsync();
        Assert.Equal(1, result);
    }

    [Fact]
    public async Task RunNonOverlappingAsync_DoesNotStartNewTask_WhenTaskRunning()
    {
        int executionCount = 0;

        async ValueTask<int> TestOverlappingTask(CancellationToken ct)
        {
            await Task.Delay(50, ct);
            Interlocked.Increment(ref executionCount);
            return 1;
        }
        
        var runner = new NonOverlappingAsyncRunner<int>(TestOverlappingTask);
        var task1 = runner.RunNonOverlappingAsync();
        var task2 = runner.RunNonOverlappingAsync();
        var result1 = await task1;
        var result2 = await task2;
        Assert.Equal(1, result1);
        Assert.Equal(1, result2);
        Assert.Equal(1, executionCount); // Assert that the task was only executed once
    }
    
    // Test fire and forget
    [Fact]
    public async Task RunNonOverlappingAsync_ShouldNotStartNewTask_WhenFireAndForgetIsUsed()
    {
        int executionCount = 0;

        async ValueTask<int> TestOverlappingTask(CancellationToken ct)
        {
            await Task.Delay(50, ct);
            Interlocked.Increment(ref executionCount);
            return 1;
        }
        
        var runner = new NonOverlappingAsyncRunner<int>(TestOverlappingTask);
        runner.FireAndForget();
        
        var task1 = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(500));
        var task2 = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(500));
        var result1 = await task1;
        var result2 = await task2;
        Assert.Equal(1, result1);
        Assert.Equal(1, result2);
        Assert.Equal(1, executionCount); // Assert that the task was only executed once
    }

    [Fact]
    public async Task RunNonOverlappingAsync_CancelsTask_WhenProxyTimeoutReached()
    {
        var runner = new NonOverlappingAsyncRunner<int>(TestTask80);
        await Assert.ThrowsAsync<TaskCanceledException>(async () => await runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(5)));
        await runner.StopAsync();
    }

    [Fact]
    public async Task RunNonOverlappingAsync_CancelsTask_WhenProxyCancellationRequested()
    {
        var runner = new NonOverlappingAsyncRunner<int>(TestTask30);
        var cts = new CancellationTokenSource();
        var task = runner.RunNonOverlappingAsync(default, cts.Token);
        cts.Cancel();
        await Assert.ThrowsAsync<TaskCanceledException>(async () => await task);
        await runner.StopAsync();
    }

    [Fact]
    public async Task RunNonOverlappingAsync_ThrowsExceptionForSynchronousTask_WhenTaskThrowsException()
    {
        var runner = new NonOverlappingAsyncRunner<int>(_ => throw new ArbitraryException());
        await Assert.ThrowsAsync<ArbitraryException>(async () => await runner.RunNonOverlappingAsync());
        await Assert.ThrowsAsync<ArbitraryException>(async () => await runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(500)));
        await Assert.ThrowsAsync<ArbitraryException>(async () => await runner.RunNonOverlappingAsync(default, new CancellationTokenSource().Token));
    }

    [Fact]
    public async Task RunNonOverlappingAsync_ReturnsThrownException_WhenSynchronousTaskThrowsException_AndProxyCancellationRequested()
    {
        var runner = new NonOverlappingAsyncRunner<int>(_ => throw new ArbitraryException());
        var cts = new CancellationTokenSource();
        await Assert.ThrowsAsync<ArbitraryException>(async () =>
        {
            var task = runner.RunNonOverlappingAsync(default, cts.Token);
            cts.Cancel();
            await task;
        });
    }
    private static async ValueTask<int> TestTaskWith10MsDelayedException(CancellationToken ct)
    {
        await Task.Delay(10, ct);
        throw new ArbitraryException();
    }

    [Fact]
    public async Task RunNonOverlappingAsync_ThrowsTaskCanceledException_WhenTaskThrowsDelayedException_AndProxyCancellationRequestedEarlier()
    {
        var runner = new NonOverlappingAsyncRunner<int>(TestTaskWith10MsDelayedException);
        var cts = new CancellationTokenSource();
        var task = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(500), cts.Token);
        cts.Cancel(); // Cancel before task completes
        var _ = await Assert.ThrowsAsync<TaskCanceledException>(async () => await task);
        await runner.StopAsync();
    }

    [Fact]
    public async Task RunNonOverlappingAsync_Exception_WhenTaskThrowsException_AndTimeoutAndCancellationUsedLater()
    {
        var runner = new NonOverlappingAsyncRunner<int>(TestTaskWith10MsDelayedException);
        var cts = new CancellationTokenSource();
        var task = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(1000), cts.Token);
        await Task.Delay(500);
        cts.Cancel();
        await Assert.ThrowsAsync<ArbitraryException>(async () => await task);
    }
    
    // Test repeated calls of Run like 5 times, including reusing after cancellation and exception
    [Fact]
    public async Task RunNonOverlappingAsync_MultipleTimesAfterCancelAndExceptions()
    {
        int executionCount = 0;
        int throwNextCount = 0;
        
        var runner = new NonOverlappingAsyncRunner<int>(async ct =>
        {
            await Task.Delay(50, ct);
            if (ct.IsCancellationRequested) throw new TaskCanceledException();
            Interlocked.Increment(ref executionCount);
            if (throwNextCount > 0)
            {
                throwNextCount--;
                throw new ArbitraryException();
            }
            return 1;
        });
        var cts = new CancellationTokenSource();
        // try with task exception first
        throwNextCount = 1;
        var task = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(1000), cts.Token);
        await Assert.ThrowsAsync<ArbitraryException>(async () => await task);
        Assert.Equal(1, executionCount);
        // Now try again, timeout being first (never guaranteed, the scheduler may be busy)
        task = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(5), cts.Token);
        await Assert.ThrowsAsync<TaskCanceledException>(async () => await task);
        Assert.Equal(1, executionCount);
        // Now try again with cancellation being first
        task = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(1000), cts.Token);
        cts.Cancel();
        await Assert.ThrowsAsync<TaskCanceledException>(async () => await task);
        Assert.Equal(1, executionCount);
        // Now try again with exception being first
        throwNextCount = 1;
        task = runner.RunNonOverlappingAsync(TimeSpan.FromMilliseconds(1000), default);
        await Assert.ThrowsAsync<ArbitraryException>(async () => await task);
        Assert.Equal(2, executionCount);
        // Now try again spamming, no exceptions
        for (int i = 0; i < 5; i++)
        {
            task = runner.RunNonOverlappingAsync(default, default);
            await task;
            Assert.Equal(3 + i, executionCount);
        }
        executionCount = 0;
        // Now try fire and forget 10x, expecting only one execution
        for (int i = 0; i < 10; i++)
        {
            runner.FireAndForget();
        }
        await runner.RunNonOverlappingAsync(default, default);
        Assert.Equal(1, executionCount);
        
        runner.Dispose();
    }
    
    // Now test dispose
    [Fact]
    public async Task Dispose_StopsTask_WhenTaskIsRunning()
    {
        int completionCount = 0;
        int startedCount = 0;
        int cancelledCount = 0;
        async ValueTask<int> TestOverlappingTask(CancellationToken ct)
        {
            try
            {
                Interlocked.Increment(ref startedCount);
                await Task.Delay(50, ct);
                Interlocked.Increment(ref completionCount);
                return 1;
            }catch(TaskCanceledException)
            {
                Interlocked.Increment(ref cancelledCount);
                return 0;
            }
        }
        
        var runner = new NonOverlappingAsyncRunner<int>(TestOverlappingTask, true);
        Assert.Equal(1, await runner.RunNonOverlappingAsync());
        Assert.Equal(1, completionCount); // Assert that the task was only executed once
        // fire and forget - may not have started
        runner.FireAndForget();
        Assert.Equal(2, startedCount);
 
        runner.Dispose();
        
        Assert.True(cancelledCount == 1 || completionCount == 2);
    }
    
    // test dispose with the last task being cancelled
    [Fact]
    public async Task Dispose_StopsTask_WhenTaskIsRunning_AndTaskIsCancelled()
    {
        
        var runner = new NonOverlappingAsyncRunner<int>(async ct =>
        {
            await Task.Delay(1, ct);
            throw new TaskCanceledException();
        });
        try
        {
            await runner.RunNonOverlappingAsync();
        }
        catch (TaskCanceledException)
        {
            // ignored
        }
        
        await runner.StopAsync();
    }
    
    // We want to do a lot of parallel testing, parallel calls to RunNonOverlappingAsync, and FireAndForget
    // And we want to StopAsync
    // and verify that all tasks are stopped
    [Fact]
    public async Task Dispose_StopsAllTasks_WhenMultipleTasksAreRunning()
    {
        int completionCount = 0;
        int startedCount = 0;
        int cancelledCount = 0;
        async ValueTask<int> TestOverlappingTask(CancellationToken ct)
        {
            try
            {
                Interlocked.Increment(ref startedCount);
                await Task.Delay(50, ct);
                Interlocked.Increment(ref completionCount);
                return 1;
            }catch(TaskCanceledException)
            {
                Interlocked.Increment(ref cancelledCount);
                throw;
            }
        }
        
        var runner = new NonOverlappingAsyncRunner<int>(TestOverlappingTask);
        var tasks = new Task[100];
        for (int i = 0; i < tasks.Length; i++)
        {
            tasks[i] = runner.RunNonOverlappingAsync().AsTask();
        }
        await runner.StopAsync();
        // ensure all proxy tasks are cancelled
        await Assert.ThrowsAsync<TaskCanceledException>(async () => await Task.WhenAll(tasks));
    }

    
}